Remove x,y,width,height from HTMLImageElement.
This conflicts with adding them to HTMLElement for custom layout.
I grepped the codebase and couldn't find any uses of these. The
big thing we lose here is a way to set width/height on an image
that automatically maintains aspect ratio, but that's something
we should move into a custom layout method anyways.

R=eseidel@chromium.org

Review URL: https://codereview.chromium.org/1060443002